chore: 코드 임시 분리
